refactored utilities controller
